# api/AlloBoulot/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation de l'API Laravel",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/companies/{companyId}/job-offers": {
            "get": {
                "tags": [
                    "Job Offers"
                ],
                "summary": "Liste des offres d'une entreprise",
                "description": "Récupère toutes les offres d'emploi d'une entreprise spécifique",
                "operationId": "8ffb58a77859f7656eb2a55bd06903e4",
                "parameters": [
                    {
                        "name": "companyId",
                        "in": "path",
                        "description": "ID de l'entreprise",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des offres de l'entreprise récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Développeur PHP Senior"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Nous recherchons..."
                                                    },
                                                    "contract_type": {
                                                        "type": "string",
                                                        "example": "CDI"
                                                    },
                                                    "study_level": {
                                                        "type": "string",
                                                        "example": "Bac+5"
                                                    },
                                                    "end_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2025-12-31"
                                                    },
                                                    "category_id": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Entreprise introuvable"
                    }
                }
            },
            "post": {
                "tags": [
                    "Job Offers"
                ],
                "summary": "Créer une offre d'emploi",
                "description": "Crée une nouvelle offre d'emploi pour une entreprise (RH/Admin uniquement)",
                "operationId": "4f650696cab023ed621c432c5bc01d25",
                "parameters": [
                    {
                        "name": "companyId",
                        "in": "path",
                        "description": "ID de l'entreprise",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Développeur PHP Senior"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Nous recherchons un développeur PHP expérimenté avec minimum 5 ans d'expérience..."
                                    },
                                    "contract_type": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "CDI"
                                    },
                                    "study_level": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Bac+5"
                                    },
                                    "end_at": {
                                        "description": "Date de fin de l'offre",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-31"
                                    },
                                    "category_id": {
                                        "description": "ID de la catégorie",
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Offre d'emploi créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Offre d'emploi créée avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Développeur PHP Senior"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Nous recherchons..."
                                                },
                                                "contract_type": {
                                                    "type": "string",
                                                    "example": "CDI"
                                                },
                                                "company_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "404": {
                        "description": "Entreprise introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/job-offers/search": {
            "get": {
                "tags": [
                    "Job Offers"
                ],
                "summary": "Rechercher des offres d'emploi",
                "description": "Recherche des offres d'emploi par mots-clés et/ou catégorie",
                "operationId": "639b6764757ab60a767b8da410fc092a",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Mot-clé de recherche (titre, description)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 255,
                            "example": "développeur php"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "ID de la catégorie",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Résultats de recherche",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Développeur PHP Senior"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Nous recherchons..."
                                                    },
                                                    "contract_type": {
                                                        "type": "string",
                                                        "example": "CDI"
                                                    },
                                                    "company_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "category_id": {
                                                        "type": "integer",
                                                        "example": 3
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entrez votre token JWT",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}